[WRFE-10](refactor): 폴더 구조 변경(page 폴더 제거)
